requirements.txt
## Necesarios
streamlit>=1.36,<2.0
tensorflow>=2.15,<2.17
numpy>=1.23,<2.0
pillow>=10.0,<12.0
opencv-python-headless>=4.8,<5.0
protobuf>=3.20.3,<5
typing-extensions>=4.8,<5


